+++ server/controller/auth/findId.js
@@ -1,0 +1,6 @@
+const { user } = require('../models');
+
+module.exports = async (id) => {
+  const found = await user.findOne({ where: id });
+  return found;
+};
