+++ server/controller/index.js
@@ -11,4 +11,21 @@ module.exports = {
   oauthController: {
     googleLogin: require('./oauth/googleLogin'),
   },
+  authController: {
+    deleteUser: require('./auth/deleteUser'),
+    findId: require('./auth/findId'),
+    login: require('./auth/login'),
+    signout: require('./auth/signout'),
+    signup: require('./auth/signup'),
+  },
+  postController: {
+    deletePost: require('./post/deletePost'),
+    editPost: require('./post/editPost'),
+    getPost: require('./post/getPost'),
+    uploadPost: require('./post/uploadPost'),
+  },
+  userController: {
+    editUserInfo: require('./user/editUserInfo'),
+    getUserInfo: require('./user/getUserInfo'),
+  },
 };
