+++ server/middleware/index.js
@@ -2,4 +2,7 @@ module.exports = {
   isAuth: require('./isAuth'),
   isLoggedIn: require('./isLoggedIn'),
   logger: require('./logger'),
+  isValidEmail: require('./isValidEmail'),
+  isValidNickname: require('./isValidNickname'),
+  lengthPassword: require('./lengthPassword'),
 };
