+++ server/router/user.js
@@ -1,6 +1,6 @@
 const express = require('express');
 
-const userController = require('../controller/user');
+const { userController } = require('../controller/index');
 
 const router = express.Router();
